// src/common/enums/post.enum.ts
export enum AllowCommentsEnum{
    allow = "allow",
    disallow = "disallow"
}

export enum AvailabilityEnum{
    public = "PUBLIC",
    friends = "FRIENDS",
    onlyMe = "ONLY_ME"
}

export enum likeActionEnum{
    like = "like",
    unlike = "unlike",
}

export enum saveActionEnum{
    save = "save",
    unsave = "unsave",
}